Remove Explicit Mapping in favour of AutoMapper
