+++ src/SFA.DAS.AssessorService.Web/StartupConfiguration/MappingStartup.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
 using SFA.DAS.AssessorService.Api.Types.Models;
+using SFA.DAS.AssessorService.Api.Types.Models.Standards;
 using SFA.DAS.AssessorService.Web.AutoMapperProfiles;
 using SFA.DAS.AssessorService.Web.Controllers.ManageUsers.ViewModels;
 using SFA.DAS.AssessorService.Web.ViewModels.Organisation;
 using SFA.DAS.AssessorService.Web.ViewModels.Search;
+using SFA.DAS.AssessorService.Web.ViewModels.Shared;
 
 namespace SFA.DAS.AssessorService.Web.StartupConfiguration
 {
@@ -16,6 +18,7 @@ namespace SFA.DAS.AssessorService.Web.StartupConfiguration
             Mapper.Initialize(cfg =>
             {
                 cfg.CreateMap<ResultViewModel, SearchResult>();
+                cfg.CreateMap<StandardVersionViewModel, StandardVersion>();
                 cfg.CreateMap<OrganisationSearchViewModel, RequestAccessOrgSearchViewModel>()
                 .ForMember(dest => dest.Address, x => x.Ignore())
                 .ForMember(dest => dest.CompanyNumber, x => x.Ignore())
