+++ src/SFA.DAS.AssessorService.Web/ViewModels/Shared/StandardVersionViewModel.cs
@@ -1,0 +1,11 @@
+﻿using SFA.DAS.AssessorService.Api.Types.Models.Standards;
+
+namespace SFA.DAS.AssessorService.Web.ViewModels.Shared
+{
+    public class StandardVersionViewModel
+    {
+        public string Title { get; set; }
+        public string StandardUId { get; set; }
+        public string Version { get; set; }
+    }
+}
